FESTOU-203 - Finalizar tela "Minhas notificações"
